+++ test/test_c.c
@@ -236,7 +236,8 @@ bool decode (uint32_t mysize, float drop_prob, uint8_t overhead)
 	// whole decoded data.
 	// note: the length of the decoded data is in bytes and might not fit
 	// a whole uint32_t.
-	uint64_t decoded_size = ceil (RaptorQ_bytes (dec) / sizeof(uint32_t));
+	uint64_t decoded_size =(uint64_t) (ceil (RaptorQ_bytes (dec)
+														/ sizeof(uint32_t)));
 	uint32_t *received = (uint32_t *) malloc (decoded_size * sizeof(uint32_t));
 
 	for (uint32_t *shit = received; shit != received + decoded_size; ++shit)
